Implement radix_array, a value-oriented sparse array
This presents an interface very similar to that of std::array, with a
few tweaks to support sparseness.  Node compression and
zero-allocation optimization are implemented.  Locking is partially
implemented (lock state is propagated, but there's currently no way to
acquire a lock to begin with).

This builds in both xv6 kernel and Linux user space (though xv6
doesn't provide std::allocator, so you can't use the default template
arguments).