Fix lwIP compilation
Unsurprisingly, I broke this with my transition to more C++-y
spinlocks and condvars.  The damage was minor but obscure: lwip/sys.h
includes spinlock.h, which is fine in C code because the header is C
compatible, but not okay in C++ code because net/sys_arch.cc includes
lwip/sys.h in an extern "C" block and spinlock.h includes atomic,
which uses templates, which can't have C linkage.