Pass current CPU to inittls and inittls the BSP much earlier
Make cpus[0] the BSP by convention so we can inittls the BSP before we
have the CPU table.  This is useful because all sorts of things crash
without tls.  On the APs, pass the CPU we're booting in a global
variable so we don't have to figure it out (which means we can do it
without a configured LAPIC).

Since cpus[0] is now required to be the BSP, remove mpbcpu() and fix
up callers.