xapic: Initialize xAPIC LAPICs using CPUID and MSRs
Previously we hard-coded a lot of things about the LAPIC.  This now
detects it using CPUID and MSRs.

Also, this moves initlapic before initcpus so we can get the LAPIC ID
of the BSP and match it up with the right CPU in the MADT.

No x2APIC support yet, but the per-CPU LAPIC object implements an
abstract interface, which the x2APIC can provide an alternate
implementation for.