ioapic: Configure IOAPIC from MADT
This introduces an abstract class for hardware IRQ-managing PICs and
implements it for the IOAPIC (the 8259A could also implement it,
though we don't bother).  Through this, the IOAPIC is configured by
ACPI, including support for multiple IOAPICs (necessary on tom and
ben), interrupt source overrides, and different polarities and
triggering modes.

This big missing piece here is support for mapping PCI interrupts.  We
use a hack that's equivalent to how we used to do this and somehow
works in QEMU and on josmp.  To do this right, we need more ACPI.