vm: Switch from VMAs to VM descriptors
A VM descriptor records the metadata for a mapping, but not where that
mapping is.  It is also small, so we can store it directly for each
mapped page (rather than pointing to a shared reference-counted VMA)
and designed to be radix node compressable for large mappings.

This switch brings with it several other changes:
* It eliminates the vmnode's page array, instead embedding the
  canonical page table in the radix array storing the VM descriptors.
  Hence, there are no more limitations on the size of a mapping.
* Reference counts are tracked per page instead of per mapping, and
  use the refcache.  As a result, pages no longer leak when mappings
  are partially remapped, and we will no longer hit a 64 core limit
  because of the size of the reference counting metadata.
* COW is performed per page instead of per mapping.
* File mappings are lazily faulted.
* Files can only be mapped in whole page increments.  As a result,
  exec must specially handle zero-padded ELF segments by eagerly
  loading the partial page on the boundary between the file-backed
  part of the segment and the zeroed part of the segment.
* sbrk now unmaps for negative n and doesn't need a buffer page
  between the break and the next mapping.

This currently uses page-level locking for read operations.  This
needs to be switched to seqlocks for consistent reads of vmdescs plus
RCU for freeing radix nodes and external elements.

This also doesn't touch the TLB code.  You still have the choice
between broadcasts and incoherent per-thread page tables.  It does set
us up to track page/core mappings.

usertests pass.  There is one known bug: since file mappings are now
lazily faulted, it's possible for us to attempt to sleep in
vmap::ensure_page -> readi while holding a spinlock (e.g., within a
page fault).