sh: Implement standardish -c argument convention
If passed -c, the following arguments are treated as a shell line to
be parsed and interpreted.  Without -c, the argument is the name of a
shell script to interpret.

This distinction is very important for supporting #! in shell scripts.
With #!, exec invokes the interpreter, passing the name of the file as
the first argument.  If sh simply treated this as a command line to
interpret, it would immediately re-exec the shell script, going into
an infinite loop.