acpi: Handle "end tag" resource type
Apparently this is how ACPI terminates resource lists, even though it
already has two *other* mechanisms for terminating lists.  Previously,
I looked for a 0-length entry to terminate the list, but this would
walk off the end of the list.  It worked before because the allocation
was likely to be followed by zeroed memory and broke when I enabled
kalloc debugging.