Use KBASE mapping in p2v whenever possible
Previously, p2v used the KCODE mapping for addresses below 2GB and the
KBASE mapping for addresses above 2GB.  This wreaked havoc with code
that was trying to iterate over memory when it hit the 2GB boundary.

Now, p2v uses KCODE until we set a flag indicating that the KBASE
mapping is set up, which causes p2v to switch to the KBASE mapping for
all translations.