kalloc: More intelligent page_info_array allocation
Previously, if the first memory hole wasn't large enough for the
page_info_array, we would dump memory until we found a hole large
enough.  On ben, this meant we wouldn't use the bottom 4GB of RAM.
Now, if we can't allocate in the first hole, we find a hole large
enough, but size the array to track all pages, not just those
following the array.