zalloc: Use a custom page list
kmem is about to become not a page list, so it won't suffice for
zalloc.  zalloc is the only kmem user outside of kalloc.  This
requires a bit of code restructuring because the free lists are now
accessed without any synchronization or atomic operations, and hence
must *only* be accessed by the local core.  Previously, we relied on
kmem's concurrency to allow for remote manipulation.