Replace argcheck{str,ptr} with copying via userptr
argcheckstr/ptr are unsafe in the presence of shared memory.
Furthermore, the technique they used of calling pagefault directly to
test for mappings and fault them if not present is going to be
inefficient with the new address space abstraction, which doesn't
provide a quick way to test the page table because there is no quick
way to do this test with soft-TLB-style address spaces.

The biggest change is that sys_{read,pread,write,pwrite} now allocate
kernel buffers to copy from/to and pass kernel pointers into the
underlying methods.  Only net actually cared and previously allocated
its own buffers and did the copy itself.  This is now both redundant
and wrong since the copy will fail with the kernel pointer, so we
strip out all of this and use the given pointer directly.