hwvm: Use relaxed loads and stores to PTEs
This ensures that reads and writes are atomic, but without unnecessary
serialization.  Now we only need to serialize before sending an IPI
from the x2apic because *that* is not a serializing operation, so the
target CPU might get the IPI before the sender's writes become
globally visible (writes to the xapic are serializing).