hwvm: Make pgmap a real class with methods and abstraction
Many of the hwvm functions are now thin wrappers around methods in
pgmap.  pgmap provides a particularly powerful iterator abstraction
that replaces walkpgdir.  Since the iterator can maintain state, it
caches the result of the page table walk, so when traversing a region,
it only has to re-walk when it moves to a different page table.