radix_array: Fix log2_exact power-of-two check
Previously, log2_exact tried to return ~0 if the argument wasn't a
power of two, but it was recursive, and calls further up the stack
were likely to add some positive value to the result, wrapping it
around.  Now log2_exact is iterative, so the ~0 error result can be
returned directly.