radix_array: Support non-power-of-two sized value types
Previously, radix_array used bit shifting and masking to compute all
subkeys, but that limited it to value types with sizes that were a
power of two.  Now it uses division and modulo (which the compiler may
or may not be smart enough to optimize into bit tricks) to support
arbitrary size value types.  Our *upper* nodes still have a
power-of-two fanout, so we can still use bit tricks for those in lieu
of exponentiation.

This also requires us to pad leaf_node up to NodeBytes, since it's now
possible for sizeof(T) to not to divide NodeBytes.