vm: vmdesc method to duplicate itself
Previously, vmap::copy knew how to duplicate a vmdesc for use in
another vmap (which copies *most* of the vmdesc, except its lock bit).
This gets more complicated with non-trivial trackers because the
cloned vmdesc needs an empty tracker.  This moves the cloning logic
into vmdesc, where both aspects can be handled cleanly by constructing
a new vmdesc directly that mirrors the current vmdesc, but is
appropriate for use in a new vmap.