refcache: Switch to an epoch scheme
The previous scheme, which depended on round-robin-ing a thread
through all of the CPUs, resulted in very high latency collection (on
the order of 2*10ms*#non-idle cores).  refcache now uses a
desynchronized epoch-based scheme where all cores periodically flush
their caches (currently, from the timer interrupt) and review simply
ensure that all cores have flushed between when it puts an object on
the review queue and when it reviews it.

This reduces memory reclamation latency enough that mapbench can run
on 48 cores without running out of memory.