crt: 16-byte align stack before entering main
GCC assumes and maintains a 16-byte stack alignment so it can use
aligned SSE instructions (which are faster than their unaligned
counterparts).  However, if you try to execute an aligned SSE
instruction with an unaligned address, you'll get a GPF.  The result
was that, if the number of characters in the program's command line
were just so, the program would start with a non-16-byte-aligned stack
and would crash as soon as it called printf with any arguments in XMM
registers (which were saved to the va_list using aligned SSE
instructions).