snzi: Specialize sref to track incrementing core
I forgot that SNZIs need to be dec'd on the same core they were inc'd
on.  Tweak the SNZI API to return a cookie (the core ID) that has to
be passed to the matching dec and add a template specialization of
sref for SNZI'd referenced that tracks the cookie in addition to the
pointer.