e1000: Object-oriented
Rather than having a hard-coded single instance of the e1000 state and
fixed functions to call for transmit and information, this introduces
an abstract netdev base class, makes the e1000 a subclass, moves all
of the e1000 methods and state into that subclass, and replaces the
hard-coded function calls with calls through an abstract netdev
instance.  This cleans up a lot of code because we don't have to
thread state around manually and means it would be easy to instantiate
multiple network devices (though obviously the LWIP glue would also
need to support this).