e1000: Rewrite init to support PCIe E1000
This rewrites the e1000 driver's initialization code to follow the
process given in the E1000 PCIe SDM.  I'm not sure what the old
initialization code was based on (it certainly didn't follow the E1000
SDM), but the new initialization code appears to work for both the PCI
and PCIe E1000's.

Something is broken on ben, though, since we don't receive interrupts
from its E1000e (though things work---slowly---if we poll).