x2apic: Unmask performance counter LVT after enabling x2APIC
This fixes profiling on ben.

Previously, we unmasked the performance counter LVT before setting the
software APIC enable bit in the spurious vector register.  On ben,
core 0's APIC was already enabled, so this was fine, but the other
cores weren't, so the unmasking didn't have an effect and this LVT
remained masked after we enabled the APIC.