lib: Standard fprintf, reconcile FILE, stdout, stderr definitions
xv6's fprintf has traditionally taken a file descriptor instead of a
FILE* (which we didn't have until recently).  Unfortunately, this,
combined with various compatibility hacks have led us to have multiple
incompatible definitions of FILE and stdout.  This patch reconciles
our FILE and stdout definitions, adds a stderr definition, changes
fprintf to take a FILE* so that it remains compatible with the
reconciled stdout definition, and introduces fdprintf, which is like
the old fprintf.

This breaks approximately everything in the known universe.  The next
patch will fix that.