trap: Fix spurious NMI panics
Because we can have two NMI sources (performance counters and
watchdog), there's a race where one overflows and triggers an NMI,
then the other overflows before we've checked for overflow and pends a
second NMI.  If this happens, we'll handle both overflows in the first
NMI and then panic when we receive the second NMI because it appears
to be spurious.

We fix this by keeping track of how many NMI sources we've processed.
If it's more than one and the next NMI happens immediately after we
return from the NMI handler, it's okay if there are no NMI sources in
the second NMI.