spinlock: Use acquire/release memory ordering
Previously, we used sequentially consistent exchanges for both acquire
and release, which resulted in an unnecessary synchronizing
instruction in release.  Now we use an acquire exchange in acquire
(which doesn't affect the generated code on x86) and a release store
in release (which turns into a regular mov on x86, as it should).