kalloc: Give each CPU its own buddy allocator
This adds a param.h option that subdivides the per-node buddy
allocators into per-CPU buddy allocators.  The span of these buddies
is unaffected (for stealing, each still spans a node and for load
balancing, each spans all of memory), so a core can return memory
allocated from any other buddy in the same node to its own buddy.

For mapbench 80 local on ben, this completely eliminates contention on
the buddy locks (which was ~40% of our time before) and brings our
performance back in line with what was in the Eurosys submission
(which makes me wonder if we weren't contending on the buddy locks for
back then for some reason).