kalloc: Free memory in to buddies steal order
Previously, we returned memory to buddies in linear order.  This is
fine if none of the buddies overlap (the order doesn't matter), but
we're about to let them overlap, so it will become important that we
return memory not just to the first buddy that contains it but to the
first buddy in the local core's steal order that contains it.