buddy: Fix bitmap allocation when tracking more than free
Previously, the buddy allocator reserved memory for the tracking
bitmaps from the tracking range, rather than the free memory range.
Originally (or without memory load balancing) there was no
distinction, but with load balancing, each buddy allocator tracks all
of memory, which means their bitmaps will all overlap in memory.

This fixes this by modifying the constructor to take both the free
range and the tracking range so that it can reserve enough bitmap
space for the entire tracking range, but it will reserve it from the
free range.