hwvm: Put tlbflush_req and have_kbase_mapping on their own cache lines
have_kbase_mapping has appeared as a hot spot on load latency
profiling and tlbflush_req is likely to be a problem in batched
shootdown mode.  Since these both have the potential to be very hot in
very different ways, isolate them both.

(Originally, we found that this significantly improved mapbench
scalability on ben, but removing scopedperf fixed the problem more and
somehow made this fix irrelevant.  But it still seems like a good
thing to do.)