kalloc: Split page_info and co-locate with NUMA nodes
Previously, we allocated all of page_info from NUMA node 0.  This
meant lots of cross-socket communication and less memory for node 0,
but had the advantage of very simple math to map between page
addresses and page_info structs.

Now, we give each NUMA node it's own page_info array and place this
array on the NUMA node.  To keep the mapping math simple and fast, we
precompute a table of page_info pointers indexed by shifted physical
address.

This has remarkably little effect on mapbench and Metis.  We still
have a huge number of reads to node 0.