refcache, vm: Statically optimize out debug messages
These are the only two places where we use the verbose print stream in
critical path code and checking the enabled flag of these streams
actually shows up in memory latency profiles.  Unfortunately, GCC
fails to inline the enabled check (in principle it could, since it
statically knows the object's type) and even if it did, some of the
print arguments can't be optimized out (calls to myid, etc).  So, wrap
all of the calls in if's that can easily be constant-folded and
eliminated.