kalloc: Use buddy free byte count instead of stats
There were various places in the load-balanced kalloc implementation
that used get_stats to get the only free byte count of a buddy
allocator, but get_stats is linear in the number of free blocks in the
allocator.  Now that the buddy allocator tracks this count directly,
use that instead.