Simpler HW=linux build
Previously we included the rules for building libu in HW=linux builds,
which meant things like ULIBA were defined, which meant we had to be
careful not to use them on HW=linux builds.  Now we simply omit
lib/Makefrag on HW=linux builds, so ULIBA and friends are simply
empty and we no longer build xv6's sysstubs.h for HW=linux builds.