vm: Hold radix lock over TLB shootdown
This fixes an important race condition where a region could be
remapped before concurrent shootdowns for that region are complete.

For example, suppose core 0 unmaps a region up to the point of sending
the shootdowns.  Previously, we wold release the lock at this point,
permitting core 1 to map that same region.  Since it was just
unmapped, mmap maps the region and returns without sending shootdowns.
However, the pages in that region may still be in the page table or
the local TLB because shootdowns from core 0 haven't completely yet.

Previously we had to release the radix lock during shootdown to
prevent deadlock because cores couldn't receive IPIs when holding the
radix lock.  However, now that the VM system only disables scheduler
preemption and not interrupts, this is safe.  It's somewhat
unfortunate that we disable preemption over TLB shootdown because it
can take as many as 40,000 cycles; however, 40,000 cycles is nothing
on the scale of the scheduler quantum, so this shouldn't cause
problems.