sh: Complete rewrite
sh is now object-oriented, cleaner, and much more flexible.  The shell
language it implements is the same except that it now supports
commands that span lines (which would have been very difficult to
support in the old code), imposes no artificial limits on line length
or argument count, and is much easier to extend.  The process creation
scheme is very different: it now creates subprocesses only for exec
and pipe, which means we can implement shell builtins very easily,
rather than hacking them in at the top level.  It also builds on
Linux and the code is even a little shorter.