refcache: Make ~referenced virtual again
My bug analysis in c834580 was incorrect (though the fix was not).  It
turns out we had made refcache::referenced::~referenced non-virtual,
which meant mnode::~mnode was non-virtual, which was causing the call
to global operator delete when we attempted to delete a monde* whose
dynamic type was a subclass of mnode.

This makes refcache::referenced::~referenced virtual again, at the
probably meaningless expense of making
refcache::referenced::referenced non-constexpr.  As a result, we don't
need to explicitly declare mnode::~mnode virtual.  It inherits its
virtual-ness from referenced, which means delete on an mnode* will use
a virtual lookup for operator delete.