radix_array: Round leaf node to power of two
Previously, we allowed the leaf nodes to contain any number of slots;
only the upper nodes needed to be a power of two.  However, this
causes the radix array to do expensive divisions frequently, so
instead round the leaf node size to a power of two and use bit
arithmetic for everything.  This wastes more memory, but improves the
performance of vm::copy in mailbench 1 1 1000 1 0 from ~600,000 cycles
to ~190,000 cycles and about halves its acquire/release time (which is
still ~350,000 cycles).