gc: Fix GC thread wakeup logic
Previously, we would wake up the GC thread whenever our local
delayed-free count was a multiple of the batch size, but this had two
problems: if the delayed-free count didn't change, we could wake up
repeatedly with no changes (this was particularly a problem if ndelay
happened to be zero and stay zero, like in linkbench), and if the
count went up by more than one in an epoch, we could skip over a
legitimate wakeup.  Fix this by instead tracking the last delayed-free
count at which we woke up the GC and wake it up if the different
between the current count and the last count exceeds the batch size.