Link binaries using g++
Previously, we linked binaries on HW=linux using g++ (to get all of
the necessary libraries), but used ld for other targets so we could
control everything.  However, some arguments, like --start-group,
require prefixing in one, but can't be used with a prefix in the
other.  So, use g++ in all cases and use -nostdlib to suppress the
things we don't want the g++ link driver to do.