sockets: Split file subclasses for LWIP and UNIX sockets
Previously these were combined into one messy subclass that behaved
very differently depending on what fields were set.  This splits them
in to two subclasses, but maintains the same odd mix of supported
functions for both.  This lets us define the two subclasses in the
source files where we actually use them, and eliminates almost all of
the net* bridge functions in favor of virtual functions on file.  This
patch tries to keep most code in the same place it was before, which
leads to some awkwardness that the following patches will smooth out.