sockets: Move AF_UNIX implementation to unixsock.cc
This moves all of the AF_UNIX logic in to class file_unix_dgram so
syssocket.cc is now simple wrappers around calls to virtual methods on
file, much like sysfile.cc.  Since everything is done with virtual
methods, we no longer need RTTI to distinguish file types.  In the
process, this also cleans up some of our from-user and to-user copying
of sockaddrs.