sockets: Use a positive number for AF_UNIX
Storing a negative AF_UNIX value in the unsigned sa_family field
causes things like == comparisons to fail.  This used to be negative
because we used it as a distinguished socket value to decide if a
file_socket was an LWIP socket or an AF_UNIX socket, but that hack is
gone.