userptr: Disallow implicit conversion to uptr
It turns out this eliminated most type-safety from userptr, since it
had both a uptr constructor and an implicit conversion to uptr (it
wasn't a complete loss because the uptr constructor was explicit, but
it wasn't great).  Now the uptr conversion is explicit, so it can
still be used, but has to be explicitly requested.