mfs: Fix mnode leakage when files are quickly created and unlinked
Previously, we initialized mnodes with a link count of 0 for
convenience.  However, this isn't safe with refcache: if the file is
created and unlinked on one core within a refcache epoch, the cached
delta will be zero, so refcache will not flush the delta, so it will
not detect the zero global count, so onzero will never get called.

This changes files to be allocated with a link count of 1 and
allocation returns the object representing the held link count, rather
than the simple reference holder.  This way, if a file's link count
does reach zero, the cached delta will be non-zero.

This also improves the performance of mnode allocation by stripping
out an unnecessary mnode cache lookup.