weakcache: Dynamically allocate in constructor
Previously, we statically constructed the buckets array, but this
causes GCC to go insane and take 30 seconds to compile mnode.cc (which
contains the mnode_cache weakcache), so dynamically allocate the
buckets array instead.  Since static constructors run before
initkalloc, this will use the linear boot allocator, so it's not
constrained for size.