avoid touching file size for access to complete file pages
File read/write needs to check file size, if accessing a partial page
at the end of the file.  But for complete file pages (e.g., either a
page that's not the last, or the last page of a file where the file
size is a multiple of PGSIZE), there's no need to read file size to
access file contents.