buddy: Use blocks in fringes of tracked region
This increases the free memory available in QEMU after boot from 272
MB to 404 MB (of 512 MB).

Previously, we initialized the buddy allocators only with 16
MB-aligned blocks of 16 MB and simply discarded any memory on the
sides of the allocated region that didn't satisfy this alignment
constraint.  We did this so that all initial blocks would have a
buddy.  This wasted a huge amount of memory because not only did we
create a buddy allocator per core, but we needed a small allocation
between each one for per-buddy allocator metadata.

Now, we initialize the buddy allocator with smaller blocks to fill out
the edges of the allocated region.  We reserve space for tracking the
buddies of these blocks, even though these buddies must never be
allocated (because this memory is used for other purposes), but simply
never put these buddy blocks into the initial free list.