Untested implementation of OpLog concept
This isn't tested at all, but I'm checking it in so people can see it.
This implements the OpLog concept, but has a rather different API than
the one proposed in the submission paper.  There's a sketch of
thoughts on the paper API and differences with this API in a comment
at the bottom of oplog.hh.