mfs: Fix termination condition in readi
Previously, if we started a read at an offset past the end of a file
but still within the last page of the file, we would think there were
bytes to copy, then adjust the end position to match the length of the
file, then attempt to copy from the offset to the end, but since the
end was no less than the offset, this would memmove 2**64-x bytes,
crashing the kernel is all sorts of catastrophic ways.

Fix this by re-checking the loop condition if we adjust end.

This fixes fstest test "fs-open_pread_pdde_1".