refcache: Fix bug that allowed onzero with non-zero true count
Previously, we didn't flush zero deltas.  Remarkably, this is wrong
and we only realized this today.  Consider

          t ->
  core 0     - * +   |   - * +   |   - * +   |
       1   +       - * +       - * +       - *
  global  1 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
  true    1 2 1 1 2 1 1 2 1 1 2 1 1 2 1 1 2 1 1
  epoch   ^-----1----^-----2-----^-----3-----^

At the end of epoch 3, the object's global count will have been zero
*and stayed zero* for two epochs, even though its true count is not
zero.

The solution is simply to flush zero deltas like we would any other
delta.  The code changes are minimal, though the explanation requires
more modification.  Remarkably, the pseudocode in the comments was
actually correct (other than it forgot to clear the cache, which was
unrelated).  Unfortunately, I can't say the same about the pseudocode
in the RadixVM paper, though only one line in the paper needs to
change: instead of flushing only non-zero entries, flush all entries.

This should also make it possible for objects to start with a zero
reference count, something that wasn't safe before.

There's a potential optimization I did not implement.  When flushing a
zero delta, if the object is already dirty or has a non-zero global
count, we can probably avoid locking or modifying the object, but I
haven't convinced myself this optimization is safe with concurrent
flushes to the same object.