lseek: Fix race we thought was okay, but isn't
We thought it was okay to read fi->off twice in the optimistic part of
lseek, but that's not safe.  For example, if two concurrent lseek(-1,
SEEK_CUR) calls both compute the same new offset, then one commits
this, then the other may see that the offset is already at what it
optimistically computed and return early.

Fix this by ensuring we only read fi->off once in the optimistic part.