Fix off-by-one in __uaccess_strend
Remarkably, this has always returned a pointer to one *past* the
terminating NULL, which is not what it claims to do.  This caused
checks to fail when fetching a user string whose NULL terminator was
at USERTOP-1 (e.g., monkstats echo 1234567), though it's not the only
problem when that happens.