mem: Print free memory statistics in pages rather than bytes
Statistics from the memory allocator (especially when dealing with a lot of
memory) is easier to understand when we use a unit which is larger than a
byte, such as pages or kilobytes. However, since the memory allocator mostly
deals with memory in terms of pages, we can get more insights into how the
system is behaving by looking at the statistics in the same units. So report
the free memory statistics in pages rather than bytes, and also print a
helpful total!