acpi: don't crash when we have too many CPUs
When we have more CPUs than NCPUs supports, we end up being unable to
boot, because we find APICIDs that we don't recognize and panic. In the
case where we have more CPUs than NCPUs, we should ignore this error,
because we should be able to get along anyway. We can continue to panic
in the common case of having a high enough NCPUs, because then there's
no reason to expect this error to occur.