uart: support real hardware serial ports
The UART on bhw2 could not receive typed input, because RTS wasn't set,
which meant the other end of the serial port wouldn't transmit anything
if it respected hardware flow control -- which bhw2's hardware does,
unlike QEMU's simulated hardware.

Similarly, AUX_OUT_2 wasn't set, which meant interrupts wouldn't be
fired when data was received. Many UARTs use this auxiliary signal to
determine whether to send interrupts, which is why it is needed.

DTR does not appear to be required on bhw2, but it appears to usually
be enabled as a "best practice" to let the other end of the connection
know that a device is present, so I'm turning it on anyway.