Pass single-step mode into qemu KVM traps
When we run in qemu with KVM mode, gdb single-steps are not handled
correctly if they are interrupted by traps; the single-step flag will
be deferred until after the interrupt returns. Fix this by detecting
the case explicitly.

Only include this in debug mode, because we wouldn't want this on real
hardware.