Replace zpage with memset
zpage was implemented using an unrolled assembly-language loop. This is
not particularly fast, compared to other solutions. Our memset
implementation, which uses rep stosb, can zero a page about three times
faster on my test machine. This is known to usually be the fastest
available on modern intel processors because of ERMSB [1], and it is
appears to be used as the default for Linux in many cases [2].

 [1] https://stackoverflow.com/a/43574756
 [2] https://lkml.org/lkml/2019/9/13/45

This commit eliminates our specialized zpage implementation, because it
is neither faster than memset, nor lends itself to more comprehensible
code.

In any case, the speed of zeroing pages does not seem to significantly
affect the time it takes to allocate a page (via the anon benchmark);
there may be an impact hidden in the noise, but not obviously so.