Convert memide to be registered as a disk
The introduction of AHCI added an abstract mechanism to handle disk
registration, but then a much clunkier system (with param.h flags) was
used to switch between AHCI, IDE, and MEMIDE disks.

Switch MEMIDE to use the same disk registration system, in the hope
that we can make this situation simpler.