Add preliminary virtual file system abstraction
sv6's home-grown filesystem extends its tendrils into every corner of
the codebase, it seems. Adding an abstraction layer is the first step
to making things more modular and comprehensible, and will enable us to
support additional filesystems, like FAT32.

The abstraction layer introduced in this commit is not final, and still
requires a lot of clean-up. But it's a place to start, and appears to
be substantially stable.