Default %p format specifier to use zero-padding
The behavior of %p is platform-specific, so we can do what we want with
it, but the '0' prefix is undefined behavior. It's very useful to have
that prefix, in order to make printed addresses more readable, so turn
it on by default.

If we were to just use the '0' prefix, then we'd get all sorts of
compiler warnings, which isn't great.