Initialize cpuhz in time for all microdelay uses
Earlier refactoring moved the initialization of cpuhz into inittsc, but
this is too late for use in a couple boot steps.

This change also adds an assertion so that things fail loudly if the
order ends up wrong, rather than just having microdelay return
instantly (which often works in QEMU but not on real hardware).