Use GRUB for boot.img disk images instead of syslinux
This enables us to use more bootloader features. The actual GRUB images are
checked into the repository so that Ward can be compiled even if grub-mkimage
isn't installed on the system.

It is possible (but hopefully unlikely?) that the hardcoded offsets in these
images that hold disk sector numbers could change between GRUB versions. If that
were to happen then the "seek=" parameters to dd in the grub/boot.img and
grub/core.img build steps would have to be updated.